nixos/minecraft-server: don't hardcode exec path
Previously, the module hard-coded a path to a 'minecraft-server'
executable script, requiring any packages that are intended to be run as
a minecraft server to name their start script in a similar manner.
With this change, packages such as papermc and purpur can name their
scripts differently as desired.
Furthermore, this means that the module will respect meta.mainProgram.